post_save_hook: Fix permissions, symlink notebook to notebooks directory
Each post save script is run as the user that is trying to save it. Previously,
everything used to be done by a single user. These saves or directory creations
are by default made with a permissions mask that disallows other users from
editing them. This was a problem for shared directories. This commit fixes that.