Set default autosave interval for notebooks to 5 seconds
My read has been that some Jupyterlab settings need to be configured
on a per-user basis. autosaveInterval depends on docmanager component.
This commit adds a default configuration for docmanager, and a
Jupyterhub pre-spawn hook that will copy this to the user's configuration
directory upon spawn. It will not overwrite if user has themselves
modified it. If the pipalhub admin wants to change this value or add
other configuration, the docmanager settings file can be found at
`etc/jupyter/lab/docmanager.jupyterlab-settings`.