Fix #7: Add on_save_notebook to dashboard service, remove build container
This commit adds to the dashboard_service to execute some code when
an event of type save-notebook is received. For now, this runs the
build.py script that will re-generate the summary page associated
with it.
It also removes build container from docker-compose.yml because it
its work is now being done by the dashboard service.