Fix dashboard_service: make default timestamp timezone-aware
The default for timestamp: datetime.utcnow() is not timezone aware.
That was causing problems when I was testing the poll.js script with
this service because browser would think it is the local time by default.
Timestamp must be timezone aware so that developers don't need to
worry about using the right way to parse.