Make setup-node.sh runnable by a sudoer
setup-node.sh could previously only be run by root. running it with
sudo would only perform all the actions (including running docker)
as the root. this isn't intended.
this commit updates the script to include sudo inside the script,
so a sudoer can run this script and run all the setup as itself.
note that if the sudoer hasn't disabled password auth, there will
be a prompt for the first time.