create-node.py now runs the installation as a non-root user
Previously create-node.py used to run setup-node.sh as root.
This commit adds a script 'create-non-root-user.sh' that takes
username as its one arg and creates a user with that username,
also making that user a sudoer.
The script also disables password authentication for that user (only
SSH, with keys copied from whatever user runs that script) so that
it can be used in scripts.