Separate install-docker.sh from setup-node.sh to fix newgrp bug
newgrp will create a subshell in which it runs with the new group.
This is okay for interactive shells, but not for our script that
we run over SSH.
Instead, the only reasonable solution seems to be to first install
and post-install docker, and then login again to run the setup-node
script.