Address #3: Add endpoint to create new users
This commit adds an endpoint to dashboard_service to create new users.
The endpoint is authenticated using HubAuth provided by jupyterhub.
This will use an Authorization header token (in the same way as JupyterHub's
REST API) to authenticate the user. The default setting is to allow
any JupyterHub user, but we can make this more specific (e.g. depending
on token scopes or if the user is an admin on JupyterHub). These extra
changes haven't been implemented in this commit because that would need
additional fiddling around in JupyterHub for a trainer, which we don't
normally do. If that has to be implemented, it should be taken on
separately.